requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
tqdm==4.60.0
pandas==1.2.4
dask==2021.4.1
pyarrow==4.0.0
jupyterlab==3.0.14
fonttools==4.22.1
Pillow==8.4.0
dvc==2.0.18
pyclipper==1.2.1
cjkwrap==2.2
kaggle==1.5.12
dvc[gs]
pytest==6.2.3
pathlib==1.0.1